feat(client,js)!: support custom jwt verifier
