+++ packages/client/rollup.config.js
@@ -1,1 +1,11 @@
-export { default } from '../../rollup.config.js';
+import defaultConfig from '../../rollup.config.js';
+
+/**
+ * @type {import('rollup').RollupOptions}
+ */
+const config = {
+  ...defaultConfig,
+  input: ['src/index.ts', 'src/shim.ts'],
+};
+
+export default config;
