+++ packages/client/src/index.test.ts
@@ -34,13 +34,6 @@ jest.mock('@logto/js', () => ({
   verifyIdToken: jest.fn(),
 }));
 
-const createRemoteJWKSet = jest.fn(async () => '');
-
-jest.mock('jose', () => ({
-  ...jest.requireActual('jose'),
-  createRemoteJWKSet: async () => createRemoteJWKSet(),
-}));
-
 describe('LogtoClient', () => {
   describe('getAccessToken', () => {
     it('should throw if idToken is empty', async () => {
@@ -230,32 +223,6 @@ describe('LogtoClient', () => {
 
       await Promise.all([logtoClient.getAccessToken(), logtoClient.getAccessToken()]);
       expect(accessTokenMap.delete).toBeCalledTimes(1);
-    });
-
-    it('should reuse jwk set', async () => {
-      requester.mockImplementation(async () => {
-        await new Promise((resolve) => {
-          setTimeout(resolve, 0);
-        });
-
-        return {
-          IdToken: 'id_token_value',
-          accessToken: 'access_token_value',
-          refreshToken: 'new_refresh_token_value',
-          expiresIn: 3600,
-        };
-      });
-
-      createRemoteJWKSet.mockClear();
-      const logtoClient = createClient(
-        undefined,
-        new MockedStorage({
-          idToken: 'id_token_value',
-          refreshToken: 'refresh_token_value',
-        })
-      );
-      await Promise.all([logtoClient.getAccessToken('a'), logtoClient.getAccessToken('b')]);
-      expect(createRemoteJWKSet).toBeCalledTimes(1);
     });
   });
