+++ packages/client/src/shim.ts
@@ -1,0 +1,25 @@
+/**
+ * @fileoverview Exports all necessary stuff for the package except the client class with default
+ * JWT verifier. It can avoid the use of `jose` package which is useful for certain environments
+ * that don't support native modules like `crypto`. (e.g. React Native)
+ */
+
+export type { IdTokenClaims, LogtoErrorCode, UserInfoResponse, InteractionMode } from '@logto/js';
+export {
+  LogtoError,
+  LogtoRequestError,
+  OidcError,
+  Prompt,
+  ReservedScope,
+  ReservedResource,
+  UserScope,
+  organizationUrnPrefix,
+  buildOrganizationUrn,
+  getOrganizationIdFromUrn,
+} from '@logto/js';
+export * from './errors.js';
+export type { Storage, StorageKey, ClientAdapter } from './adapter/index.js';
+export { PersistKey, CacheKey } from './adapter/index.js';
+export { createRequester } from './utils/index.js';
+export * from './types/index.js';
+export * from './client.js';
