+++ packages/js/package.json
@@ -31,8 +31,7 @@
   },
   "dependencies": {
     "@silverhand/essentials": "^2.8.7",
-    "camelcase-keys": "^7.0.1",
-    "jose": "^5.0.0"
+    "camelcase-keys": "^7.0.1"
   },
   "devDependencies": {
     "@silverhand/eslint-config": "^5.0.0",
@@ -45,6 +44,7 @@
     "jest": "^29.5.0",
     "jest-environment-jsdom": "^29.5.0",
     "jest-matcher-specific-error": "^1.0.0",
+    "jose": "^5.0.0",
     "lint-staged": "^15.0.0",
     "nock": "^13.3.0",
     "prettier": "^3.0.0",
